backend/requirements.txt
fastapi
uvicorn
websockets
python-dotenv
google-genai
numpy
